better defense of Python as implementation language
Speaking for myself at least, the problem with R and Weka is not
so much the statistics but the programming languages: R is a
poorly designed language, Weka uses two different interfaces
(GUI and Java).

I dropped Shogun because I'm not sure if it has the same audience
as the others. It seems to focus on kernel learning and very large
datasets.